# backend/openapi.yaml
openapi: 3.0.3
info:
  title: V PATHing Rewards API
  version: 1.0.0
  description: Cashback rewards platform API
servers:
  - url: https://vpathrewards.store/api
    description: Production
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
...
    Offer:
      type: object
      properties:
        id:
          type: integer
        merchant_id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        cashback_rate:
          type: number
        affiliate_link:
          type: string
        is_active:
          type: integer
        merchant_name:
          type: string
        merchant_logo:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
...
paths:
...
  /merchants/{id}/offers:
    get:
      operationId: getMerchantOffers
      summary: Get offers for a merchant
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Offers
...
  /offers:
    get:
      operationId: listOffers
      summary: List active offers
      security: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
            maxLength: 200
        - name: category
          in: query
          schema:
            type: string
            maxLength: 100
        - name: merchantId
          in: query
          schema:
            type: integer
            minimum: 1
        - name: minCashback
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 100
        - name: maxCashback
          in: query
          schema:
            type: number
            minimum: 0
            maximum: 100
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, cashback_asc, cashback_desc, newest]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Paginated offer list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Offer'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /offers/{id}:
    get:
      operationId: getOffer
      summary: Get offer by ID
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Offer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offer'
        '403':
          description: Not available in your state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ─── Cashback ──────────────────────────────────────────────────────────────
...